+++ src/main/java/com/ssaw/InventoryManagement/entity/SecuritiesClosedPayInventory.java
@@ -26,7 +26,7 @@ public class SecuritiesClosedPayInventory {
     //业务状态 1流入，-1流出
     private int flag;
     //总金额*
-    private BigDecimal totalPrice;
+    private Double totalPrice;
     //备注
     private String securitiesClosedPayDesc;
     //期初标志 是否从其他系统导入得期初数据 0：不是 1：是
@@ -48,7 +48,7 @@ public class SecuritiesClosedPayInventory {
      * @param securitiesClosedPayDesc   备注
      * @param securityPeriodFlag    期初标志 是否从其他系统导入得期初数据 0：不是 1：是
      */
-    public SecuritiesClosedPayInventory(String securitiesClosedPayInventoryId, String dateTime, String fundId, String securitiesId, int securitiesType, int flag, BigDecimal totalPrice, String securitiesClosedPayDesc, int securityPeriodFlag) {
+    public SecuritiesClosedPayInventory(String securitiesClosedPayInventoryId, String dateTime, String fundId, String securitiesId, int securitiesType, int flag, Double totalPrice, String securitiesClosedPayDesc, int securityPeriodFlag) {
         this.securitiesClosedPayInventoryId = securitiesClosedPayInventoryId;
         this.dateTime = dateTime;
         this.fundId = fundId;
@@ -112,11 +112,11 @@ public class SecuritiesClosedPayInventory {
         this.flag = flag;
     }
 
-    public BigDecimal getTotalPrice() {
+    public Double getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(BigDecimal totalPrice) {
+    public void setTotalPrice(Double totalPrice) {
         this.totalPrice = totalPrice;
     }
